+++ js/deprecated.js
@@ -2,11 +2,13 @@
  *
  * By Michael Teeuw http://michaelteeuw.nl
  * MIT Licensed.
+ *
+ * Olex S. original idea this deprecated option
  */
 
-var deprecated = [
-    'kioskmode'
-];
+var deprecated = {
+	configs: ["kioskmode"]
+};
 
 /*************** DO NOT EDIT THE LINE BELOW ***************/
-if (typeof module !== 'undefined') {module.exports = deprecated;}
+if (typeof module !== "undefined") {module.exports = deprecated;}